CLIP-clip/requirements.txt
numpy
scipy
pandas
torch~=1.7.1
torchvision~=0.8.2
pillow
requests
ftfy
regex
tqdm
grad-cam